requirements.txt
altgraph==0.17.4
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.34
botocore==1.35.34
branca==0.7.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dill==0.3.8
dirtree==0.3.0.dev0
filelock==3.15.4
Flask==3.0.3
folium==0.17.0
fonttools==4.53.1
fsspec==2024.6.1
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
macholib==1.16.3
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pillow==10.4.0
pipdeptree==2.23.1
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.14
PyQt5_sip==12.15.0
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.14
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
scipy==1.13.1
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
sympy==1.13.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.84
ultralytics-thop==2.0.5
urllib3==1.26.15
Werkzeug==3.0.4
xyzservices==2024.6.0
yacs==0.1.8
zipp==3.20.1
